storeの構成変更
